LayoutTests/inspector/page/setShowPaintRects-expected.txt

Tests for the Page.setShowPaintRects.


== Running test suite: Page.setShowPaintRects
-- Running test case: Page.setShowPaintRects.Enabled
PASS: Should not have paint rects displayed.
PASS: Should have paint rects displayed.
PASS: Should not have paint rects displayed.

-- Running test case: Page.setShowPaintRects.Disabled
PASS: Should not have paint rects displayed.
PASS: Should not have had any paint rects displayed.
